refactor: Move 'core/message' -> 'base/message'
- 追加予定のcore/memoryがmessageへ依存するため、一方向依存を守る
- engine/base/message.hへ変更
